requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-db-connection-pool==1.2.5
djangorestframework==3.15.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
PyMySQL==1.1.1
python-decouple==3.8
SQLAlchemy==2.0.30
sqlparams==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.0
